% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDBFunction.R
\name{ChangeCellChatDB}
\alias{ChangeCellChatDB}
\title{Transform the database into a format suitable for CellChat}
\usage{
ChangeCellChatDB(priorDatabase = NULL, extension = TRUE, species = NULL)
}
\arguments{
\item{priorDatabase}{Dataframe with 'ligand', 'receptor' and 'lr' three columns}

\item{extension}{Logic. Whether use the combination of default database of CellChat and other databases}

\item{species}{human, mouse and zebrafish}
}
\value{
list
}
\description{
Transform the database into a format suitable for CellChat
}
\author{
Jiaxin Luo
}
