% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDBFunction.R
\name{ChangeCellPhoneDB}
\alias{ChangeCellPhoneDB}
\title{Transform the database into a format suitable for CellPhoneDB}
\usage{
ChangeCellPhoneDB(priorDatabase = NULL, extension = TRUE)
}
\arguments{
\item{priorDatabase}{Dataframe with 'ligand', 'receptor' and 'lr' three columns}

\item{extension}{Logic. Whether use the combination of default database of CellPhoneDB and other databases.}
}
\value{
db_path: The folder path used to save new database
}
\description{
Transform the database into a format suitable for CellPhoneDB
}
\author{
Jiaxin Luo
}
