% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDBFunction.R
\name{ChangeDominoDB}
\alias{ChangeDominoDB}
\title{Transform the database into a format suitable for Domino}
\usage{
ChangeDominoDB(priorDatabase = NULL)
}
\arguments{
\item{priorDatabase}{Dataframe with 'ligand', 'receptor' and 'lr' three columns}
}
\value{
db_path: The folder path used to save new database
}
\description{
Transform the database into a format suitable for Domino
}
\author{
Jiaxin Luo
}
