% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDBFunction.R
\name{ChangeICELLNETDB}
\alias{ChangeICELLNETDB}
\title{Transform the database into a format suitable for ICELLNET}
\usage{
ChangeICELLNETDB(priorDatabase = NULL, extension = TRUE)
}
\arguments{
\item{priorDatabase}{Dataframe with 'ligand', 'receptor' and 'lr' three columns}

\item{extension}{Logic. Whether use the combination of default database of ICELLNET and other databases.}
}
\value{
dataframe
}
\description{
Transform the database into a format suitable for ICELLNET
}
\author{
Jiaxin Luo
}
