% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDBFunction.R
\name{ChangeNATMIDB}
\alias{ChangeNATMIDB}
\title{Transform the database into a format suitable for NATMI}
\usage{
ChangeNATMIDB(
  priorDatabase = NULL,
  extension = TRUE,
  keep_complexes = TRUE,
  NATMI_dir = NULL
)
}
\arguments{
\item{priorDatabase}{Dataframe with 'ligand', 'receptor' and 'lr' three columns}

\item{extension}{Logic. Whether use the combination of default database of NATMI and other databases}

\item{keep_complexes}{Logic. If TRUE, multi-subunit of ligand and receptor will be splited into single-subunit. Otherwise, the interactions containing multi-subunit will be filtered out.}

\item{NATMI_dir}{Folder path of source code of NATMI. If NULL, it will download the source code of NATMI from github in the current directory.}
}
\value{
NATMI_dir Folder path of source code of NATMI
}
\description{
Transform the database into a format suitable for NATMI
}
\author{
Jiaxin Luo
}
