% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDBFunction.R
\name{ChangeRNAMagnetDB}
\alias{ChangeRNAMagnetDB}
\title{Transform the database into a format suitable for RNAMagnet}
\usage{
ChangeRNAMagnetDB(priorDatabase = NULL, extension = TRUE)
}
\arguments{
\item{priorDatabase}{Dataframe with 'ligand', 'receptor' and 'lr' three columns}

\item{extension}{Logic. Whether use the combination of default database of RNAMagnet and other databases.}
}
\value{
dataframe
}
\description{
Transform the database into a format suitable for RNAMagnet
}
\author{
Jiaxin Luo
}
