% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDBFunction.R
\name{ChangescSeqCommDB}
\alias{ChangescSeqCommDB}
\title{Transform the database into a format suitable for scSeqComm}
\usage{
ChangescSeqCommDB(priorDatabase = NULL)
}
\arguments{
\item{priorDatabase}{Dataframe with 'ligand', 'receptor' and 'lr' three columns}
}
\value{
dataframe
}
\description{
Transform the database into a format suitable for scSeqComm
}
\author{
Jiaxin Luo
}
