% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDBFunction.R
\name{ChangeSCSRDB}
\alias{ChangeSCSRDB}
\title{Transform the database into a format suitable for SingleCellSignalR}
\usage{
ChangeSCSRDB(priorDatabase = NULL, extension = TRUE, keep_complexes = TRUE)
}
\arguments{
\item{priorDatabase}{Dataframe with 'ligand', 'receptor' and 'lr' three columns}

\item{extension}{Logic. Whether use the combination of default database of SingleCellSignalR and other databases}

\item{keep_complexes}{Logic. If TRUE, multi-subunit of ligand and receptor will be splited into single-subunit. Otherwise, the interactions containing multi-subunit will be filtered out.}
}
\value{
dataframe
}
\description{
Transform the database into a format suitable for SingleCellSignalR
}
\author{
Jiaxin Luo
}
